Speed up and simplify alchemy RF tests, and move test_replace_reaction_field in test_forces.py
