+++ openmmtools/tests/test_alchemy.py
@@ -27,7 +27,7 @@ import nose
 import scipy
 from nose.plugins.attrib import attr
 
-from openmmtools import testsystems
+from openmmtools import testsystems, forces
 from openmmtools.constants import kB
 from openmmtools.alchemy import *
 
@@ -41,9 +41,7 @@ logger = logging.getLogger(__name__)
 temperature = 300.0 * unit.kelvin  # reference temperature
 # MAX_DELTA = 0.01 * kB * temperature # maximum allowable deviation
 MAX_DELTA = 1.0 * kB * temperature  # maximum allowable deviation
-MAX_FORCE_RELATIVE_ERROR = 1.0e-6 # maximum allowable relative force error
 GLOBAL_ENERGY_UNIT = unit.kilojoules_per_mole  # controls printed units
-GLOBAL_FORCE_UNIT = unit.kilojoules_per_mole / unit.nanometers # controls printed units
 GLOBAL_ALCHEMY_PLATFORM = None  # This is used in every energy calculation.
 # GLOBAL_ALCHEMY_PLATFORM = openmm.Platform.getPlatformByName('OpenCL') # DEBUG: Use OpenCL over CPU platform for testing since OpenCL is deterministic, while CPU is not
 
@@ -88,52 +86,6 @@ def compute_energy(system, positions, platform=None, force_group=-1):
     return potential
 
 
-def compute_forces(system, positions, platform=None, force_group=-1):
-    """Compute forces of the system in the given positions.
-
-    Parameters
-    ----------
-    platform : simtk.openmm.Platform or None, optional
-        If None, the global GLOBAL_ALCHEMY_PLATFORM will be used.
-    force_group : int flag or set of int, optional
-        Passed to the groups argument of Context.getState().
-
-    """
-    timestep = 1.0 * unit.femtoseconds
-    integrator = openmm.VerletIntegrator(timestep)
-    context = create_context(system, integrator, platform)
-    context.setPositions(positions)
-    state = context.getState(getForces=True, groups=force_group)
-    forces = state.getForces(asNumpy=True)
-    del context, integrator, state
-    return forces
-
-
-def generate_new_positions(system, positions, platform=None, nsteps=50):
-    """Generate new positions by taking a few steps from the old positions.
-    Parameters
-    ----------
-    platform : simtk.openmm.Platform or None, optional
-        If None, the global GLOBAL_ALCHEMY_PLATFORM will be used.
-    nsteps : int, optional, default=50
-        Number of steps of dynamics to take.
-    Returns
-    -------
-    new_positions : simtk.unit.Quantity of shape [nparticles,3] with units compatible with distance
-        New positions
-    """
-    temperature = 300 * unit.kelvin
-    collision_rate = 90 / unit.picoseconds
-    timestep = 1.0 * unit.femtoseconds
-    integrator = openmm.LangevinIntegrator(temperature, collision_rate, timestep)
-    context = create_context(system, integrator, platform)
-    context.setPositions(positions)
-    integrator.step(nsteps)
-    new_positions = context.getState(getPositions=True).getPositions(asNumpy=True)
-    del context, integrator
-    return new_positions
-
-
 def minimize(system, positions, platform=None, tolerance=1.0*unit.kilocalories_per_mole/unit.angstroms, maxIterations=50):
     """Minimize the energy of the given system.
 
@@ -162,7 +114,7 @@ def minimize(system, positions, platform=None, tolerance=1.0*unit.kilocalories_p
     return minimized_positions
 
 
-def compute_energy_force(system, positions, force_name):
+def compute_force_energy(system, positions, force_name):
     """Compute the energy of the force with the given name."""
     system = copy.deepcopy(system)  # Copy to avoid modifications
     force_name_index = 1
@@ -506,40 +458,6 @@ def compare_system_energies(reference_system, alchemical_system, alchemical_regi
         raise Exception(err_msg.format(delta / unit.kilocalories_per_mole, MAX_DELTA / unit.kilocalories_per_mole))
 
 
-def compare_system_forces(reference_system, alchemical_system, positions, name="", platform=None):
-    """Check that the forces of reference and modified systems are close.
-
-    Parameters
-    ---------
-    reference_system : simtk.openmm.System
-        Reference System
-    alchemical_system : simtk.openmm.System
-        System to compare to reference
-    positions : simtk.unit.Quantity of shape [nparticles,3] with units of distance
-        The particle positions to use
-    name : str, optional, default=""
-        System name to use for debugging.
-    platform : simtk.openmm.Platform, optional, default=None
-        If specified, use this platform
-
-    """
-    # Compute forces
-    reference_force = compute_forces(reference_system, positions, platform=platform) / GLOBAL_FORCE_UNIT
-    alchemical_force = compute_forces(alchemical_system, positions, platform=platform) / GLOBAL_FORCE_UNIT
-
-    # Check that error is small.
-    def magnitude(vec):
-        return np.sqrt(np.mean(np.sum(vec**2, axis=1)))
-
-    relative_error = magnitude(alchemical_force - reference_force) / magnitude(reference_force)
-    if np.any(np.abs(relative_error) > MAX_FORCE_RELATIVE_ERROR):
-        print("========")
-        err_msg = ("Maximum allowable relative force error exceeded (was {:.8f}; allowed {:.8f}).\n"
-                   "alchemical_force = {:.8f}, reference_force = {:.8f}, difference = {:.8f}")
-        raise Exception(err_msg.format(relative_error, MAX_FORCE_RELATIVE_ERROR, magnitude(alchemical_force),
-                                       magnitude(reference_force), magnitude(alchemical_force-reference_force)))
-
-
 def check_interacting_energy_components(reference_system, alchemical_system, alchemical_regions, positions):
     """Compare full and alchemically-modified system energies by energy component.
 
@@ -676,7 +594,7 @@ def check_interacting_energy_components(reference_system, alchemical_system, alc
     for force_name in ['HarmonicBondForce', 'HarmonicAngleForce', 'PeriodicTorsionForce',
                        'GBSAOBCForce', 'CustomGBForce']:
         alchemical_forces_energies = [energy for label, energy in energy_components.items() if force_name in label]
-        reference_force_energy = compute_energy_force(reference_system, positions, force_name)
+        reference_force_energy = compute_force_energy(reference_system, positions, force_name)
 
         # There should be no force in the alchemical system if force_name is missing from the reference
         if reference_force_energy is None:
@@ -811,7 +729,7 @@ def check_noninteracting_energy_components(reference_system, alchemical_system, 
                                     if i not in alchemical_regions.alchemical_atoms]
 
         # Compute reference force energy.
-        reference_force_energy = compute_energy_force(system, non_alchemical_positions, force_name)
+        reference_force_energy = compute_force_energy(system, non_alchemical_positions, force_name)
         assert_almost_equal(reference_force_energy, alchemical_energy,
                             'reference {}, alchemical {}'.format(reference_force_energy, alchemical_energy))
 
@@ -1240,12 +1158,7 @@ class TestAbsoluteAlchemicalFactory(object):
     def generate_cases(cls):
         """Generate all test cases in cls.test_cases combinatorially."""
         cls.test_cases = dict()
-        switched_rf_factory = AbsoluteAlchemicalFactory(alchemical_rf_treatment='switched')
-        shifted_rf_factory = AbsoluteAlchemicalFactory(alchemical_rf_treatment='shifted')
-
-        # Create reference versions of all rf-containing systems with their switched counterparts with c_rf = 0
-        for (name, testsystem) in cls.test_systems.items():
-            setattr(testsystem, 'modified_rf_system', switched_rf_factory.replace_reaction_field(testsystem.system))
+        factory = AbsoluteAlchemicalFactory(alchemical_rf_treatment='switched')
 
         # We generate all possible combinations of annihilate_sterics/electrostatics
         # for each test system. We also annihilate bonds, angles and torsions every
@@ -1284,20 +1197,24 @@ class TestAbsoluteAlchemicalFactory(object):
                                              softcore_c=1.0, softcore_d=1.0, softcore_e=1.0, softcore_f=1.0)
                     test_case_name += ', modified softcore parameters'
 
-                # Also store shifted rf alchemical system for energy component comparisons
-                shifted_rf_alchemical_system = shifted_rf_factory.create_alchemical_system(test_system.system, region)
-                setattr(test_system, 'shifted_rf_alchemical_system', shifted_rf_alchemical_system)
+                # Pre-generate alchemical system.
+                alchemical_system = factory.create_alchemical_system(test_system.system, region)
 
-                # Pre-generate alchemical system with switched rf
-                alchemical_system = switched_rf_factory.create_alchemical_system(test_system.system, region)
+                # If the test system uses reaction field replace reaction field
+                # of the reference system to allow comparisons.
+                nonbonded_force = forces.find_nonbonded_force(test_system.system)
+                if nonbonded_force.getNonbondedMethod() == openmm.NonbondedForce.CutoffPeriodic:
+                    forcefactories.replace_reaction_field(test_system.system, return_copy=False,
+                                                          switch_width=factory.switch_width)
+
+                # Add test case.
                 cls.test_cases[test_case_name] = (test_system, alchemical_system, region)
-
                 n_test_cases += 1
 
     def test_fully_interacting_energy(self):
         """Compare the energies of reference and fully interacting alchemical system."""
         for test_name, (test_system, alchemical_system, alchemical_region) in self.test_cases.items():
-            f = partial(compare_system_energies, test_system.modified_rf_system,
+            f = partial(compare_system_energies, test_system.system,
                         alchemical_system, alchemical_region, test_system.positions)
             f.description = "Testing fully interacting energy of {}".format(test_name)
             yield f
@@ -1310,25 +1227,6 @@ class TestAbsoluteAlchemicalFactory(object):
             f.description = "Testing non-interacting energy of {}".format(test_name)
             yield f
 
-    def test_replace_reaction_field(self):
-        """Check that replacing reaction-field electrostatics with Custom*Force
-        yields minimal force differences with original system.
-
-        Note that we cannot test for energy consistency or energy overlap because
-        which atoms are within the cutoff will cause energy difference to vary wildly.
-
-        """
-        platform = openmm.Platform.getPlatformByName('Reference')
-        factory = AbsoluteAlchemicalFactory(alchemical_rf_treatment='switched', switch_width=None)
-        for test_name, (test_system, alchemical_system, alchemical_region) in self.test_cases.items():
-            if (test_system.system.getNumForces() != test_system.modified_rf_system.getNumForces()):
-                modified_rf_system = factory.replace_reaction_field(test_system.system)
-                # Make sure positions are not at minimum
-                positions = generate_new_positions(test_system.system, test_system.positions)
-                f = partial(compare_system_forces, test_system.system, modified_rf_system, positions, name=test_name, platform=platform)
-                f.description = "Testing replace_reaction_field on system {}".format(test_name)
-                yield f
-
     @attr('slow')
     def test_fully_interacting_energy_components(self):
         """Test interacting state energy by force component."""
@@ -1338,10 +1236,8 @@ class TestAbsoluteAlchemicalFactory(object):
                             if 'Explicit' in test_name]
         for test_name in test_cases_names:
             test_system, alchemical_system, alchemical_region = self.test_cases[test_name]
-            # We have to compare shifted rf system with original system because test cannot handle
-            # re-coded reaction-field forces with c_rf = 0
-            f = partial(check_interacting_energy_components, test_system.system, test_system.shifted_rf_alchemical_system,
-                    alchemical_region, test_system.positions)
+            f = partial(check_interacting_energy_components, test_system.system, alchemical_system,
+                        alchemical_region, test_system.positions)
             f.description = "Testing energy components of %s..." % test_name
             yield f
 
@@ -1363,10 +1259,12 @@ class TestAbsoluteAlchemicalFactory(object):
         for platform in platforms:
             GLOBAL_ALCHEMY_PLATFORM = platform
             for test_name, (test_system, alchemical_system, alchemical_region) in self.test_cases.items():
-                f = partial(compare_system_energies, test_system.modified_rf_system, alchemical_system, alchemical_region, test_system.positions)
+                f = partial(compare_system_energies, test_system.system, alchemical_system,
+                            alchemical_region, test_system.positions)
                 f.description = "Test fully interacting energy of {} on {}".format(test_name, platform.getName())
                 yield f
-                f = partial(check_noninteracting_energy_components, alchemical_system, alchemical_region, test_system.positions)
+                f = partial(check_noninteracting_energy_components, test_system.system, alchemical_system,
+                            alchemical_region, test_system.positions)
                 f.description = "Test non-interacting energy of {} on {}".format(test_name, platform.getName())
                 yield f
 
@@ -1380,7 +1278,7 @@ class TestAbsoluteAlchemicalFactory(object):
             #cached_trajectory_filename = os.path.join(os.environ['HOME'], '.cache', 'alchemy', 'tests',
             #                                           test_name + '.pickle')
             cached_trajectory_filename = None
-            f = partial(overlap_check, test_system.modified_rf_system, alchemical_system, test_system.positions,
+            f = partial(overlap_check, test_system.system, alchemical_system, test_system.positions,
                         cached_trajectory_filename=cached_trajectory_filename, name=test_name)
             f.description = "Testing reference/alchemical overlap for {}".format(test_name)
             yield f
